share/noba-web/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.1
psutil>=5.9.8
pydantic>=2.9.0
pydantic-settings>=2.5.0
pyjwt>=2.8.0
passlib[bcrypt]>=1.7.4
pyyaml>=6.0.1
